+++ providers/sqlite/CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+## 0.3.20
+
 ## 0.3.19
 
 ### Features